Mark pods as not ready when host goes offline
When a host goes offline unexpectedly (e.g. spontaneous reboot or
power failure), it takes a while before kubernetes detects that
the node is unavailable and sets the node to be not ready, which
causes the pods on the node to be set to not ready as well. Until
that happens, the endpoints for the pods are still in use, which
can lead to message loss.

In StarlingX, our maintenance component detects that a node is
offline faster than kubernetes does. To improve the time taken
to react, the VIM is updated to set the status of the pods on
an offline node to not ready when it is informed by maintenance
that a node is offline.

Change-Id: I1a20707981c5517e0d305fd9551bc5e69c0282cd
Story: 2002843
Task: 33695
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>