Correct VIM host audit criteria for failing instances
The VIM's host audit will fail instances on any host that is
"not enabled". That includes hosts where the operational state
is unknown.

Updating the check to ensure the host is "disabled" not that it
is "not enabled" to avoid failing instances on a host where we
don't know the operational state.

Change-Id: I68d3e9f63695de721c10fb1dd2b7ac5917cb50fa
Closes-Bug: 1840176
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>