Make VIM tolerant of compute service failures
When the VIM detects the nova compute service is down on a
worker host, it attempts to migrate instances off that host
(by "disabling" the host). However, this isn't possible if
the compute service is down. The VIM then fails the instances,
which will eventually result in their evacuation (if the host
goes offline) or a reboot of the instance (if the compute
service recovers).

In the containers world, when the libvirt pod is restarted
(e.g. when stx-openstack application is re-applied), nova
reports that the compute service is down (for a short period
of time), which causes the undesirable behaviour described
above. The VIM is being updated to not disable the host in
this case and instead just raise an alarm to indicate that
the compute service has failed.

Change-Id: I186d8d76bbcd87405bafec47deb92ec24580640e
Closes-Bug: 1833096
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>