Wait for pods to terminate when host is locked
When a host is locked, the VIM adds the NoExecute taint to the
host, which causes pods on the host to be terminated. The VIM
is currently not waiting for the pods to terminate, so if the
host reboots before pods are terminated (e.g. if an unlock was
done quickly after the lock), there can be issues with pod
recovery.

With this change, the VIM will wait for up to 3 minutes for
pods to terminate. If there are still pods terminating at
that point, the VIM will either fail the lock (normal case)
or allow the lock to pass (force lock case).

Change-Id: I9f402ab6dc29ecc314bbd1a058ae3d79ea8728fe
Closes-Bug: 1846818
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>