Fix up nfv tox targets for python3
basepython for tox.ini in nfv is not specified
so it varies when running on different zuul machines.

This fixes bandit (which is not enabled in zuul) to use
alternative path declaration for excluded folders.
See: https://github.com/PyCQA/bandit/issues/400

This fixes pylint when run in python3 since the
suppression syntax is different in both versions
(pylint: disable) vs (pylint disable)
https://github.com/PyCQA/pylint/issues/986

This fixes a stricter check that was introduced by
pylint 2.4.3 (released Oct 19, 2019) related to
abs.abstractproperty
https://github.com/PyCQA/pylint/issues/3150

Change-Id: I3718229d6d0e6074cf10347e6c6cb40fb26c25db
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>