Ignore instance action for non-OpenStack worker hosts
When a patch orchestration strategy is created by the dcmanager,
the default instance action is set to migrate to minimize the
impact to any OpenStack instances (VMs) that may be running.
However, on simplex subclouds, the migrate instance action is
rejected, since there is nowhere to migrate these instances.
The user can override this default, but for non-OpenStack
subclouds, the default instance action is not applicable and
should be ignored by the VIM so the user does not need to
change it.

Updating the VIM to ignore the instance action for simplex hosts
that do not have OpenStack installed.

Change-Id: I70984b1042f4460fd8a8e621e44756f022df03d7
Closes-Bug: 1854655
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>