Allow VIM to shutdown when all FDs are used
When the VIM attempts to shut down because it runs out of FDs,
the shutdown code attempts to create a file, before it activates
a SIGALRM timer that triggers fail safe shutdown. When all FDs
have been consumed, the shutdown code hangs while creating the
file and never shuts down.

This is fixed by changing the shutdown code to first start the
failsafe SIGALRM timer. That way the process is guaranteed to
be terminated even if it hangs while attempting to create the
file.

Change-Id: I7184fe1333d2641227c193f7c8e2e7b1bebcfed8
Partial-Bug: 1862049
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>