Add REST API timeouts to nfv-client
The VIM's nfv-client sends REST API messages to the VIM, but it
currently does not set a timeout for these calls. In the case
where a message is lost, the nfv-client would block forever,
causing the calling process to hang.

Fixing this by adding timeouts to the nfv-client to ensure that
all REST API messages it sends will timeout even in the case of
lost messages.

Change-Id: Iba06924fb7bd14a1ee3362b4fd19aa4114dc34cd
Partial-Bug: 1886037
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>