Fix the exception of taint/untaint node operation in VIM
If Kubernetes node has images without tag and digest
("crictl images" shows <none>:<none>), taint/untaint node
operations in VIM(triggerd by host lock/unlock) that send
the requests to Kubernetes to read/patch node via python
Kubernetes client will raise this exception.
  "Invalid value for `names`, must not be `None`"

The exception is raised in python Kubernetes client but not
from kubernetes. The node is actually patched successfully.

Implement a workaround to replace V1ContainerImage.names
in the python Kubernetes in the python Kubernete client
to bypass the "none image" check.

This workaround should be removed if the proposed solutions
can be made in kubernetes or a workaround can be implemented
in containerd.
https://github.com/kubernetes/kubernetes/pull/79018
https://github.com/containerd/containerd/issues/4771

Closes-Bug: 1905481
Signed-off-by: Angie Wang <angie.wang@windriver.com>
Change-Id: I747b5309747007204033d8dc2a850d3f824ddc5d