Reset alarm timers back to 30 after orch sets them to 2
The alarm audit is set to 2 seconds by an alarm callback.
Once that callback has been processed, need to reset the
timer back to its default of 30 seconds.

Note: Patch orch did not encounter this issue, because its
audit called other code after the alarm callback, which reset
the timer.

Closes-Bug: 1921127
Change-Id: Iba82c71757c0a9e760926e81a42ba91b14196e0e
Signed-off-by: albailey <Al.Bailey@windriver.com>