Adding parallel apply capabilities to kube orch
When a kubernetes orch defines a parallel apply type for workers
or storage, the code used by patch orchestration to group the
hosts, and add appropriate steps is now being re-purposed during
this phase of kube orch.

This change moves three of the patch host stage methods into mixins
so that the code can be reused among different strategy types.

The kubelet stages are now similar to patch in how they handle
parallel apply types and VM hosts.

The action type can now be passed in, and defaults to stop-start
instead of migrate.

Story: 2008137
Task: 42198
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: I284c84dea962c1d8f5548472767f5f9b3c6325d7