Handle patch install failure scenario in patch orchestration
When a patch installs an RPM, but there is a post-installation failure
caused by a restart script error, the host is considered both patch current
and also patch failed.

However VIM patch orchestration was considering it as successful.
The vim logic to check the patch-failed flag is now being done before
the patch current check, so properly characterize this type
of patch orchestration problem.

Closes-Bug: 1842952
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: I2f8a784be4702537abff7996156344a3b558aefe