Add missing debug logger entries for VIM
The debug loggers are pre-created during process startup,
and any that are missing can be created using lazy init.

However the lazy init is not multi process safe and can
lead to a subprocess deadlock during the import phase, which
produces no logs or evidence that the thread is stalled.

The workaround at the moment is to ensure that all the debug
loggers defined at import time are configured to be setup when
the process is created.

There is still another cause of a stall. A separate commit
will address it.

Partial-Bug: 1925697
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: I90436c1bd063a634ab8a1496236a78b01a04d51a