Update fw-update-strategy steps to load wait_time from_dict
The sw-manager fw-update-strategy feature is seen
to fail in a traceback.

The __wait_time member of the FwUpdateHostsStep and
FwUpdateAbortHostsStep objects are not de-serialized
from the DB using the ‘from_dict’ methods. This means
it does not run the ‘init’ method for those classes,
but instead attempts to re-constitute the object
directly which can lead to an exception\traceback.

This update adds the _wait_time member to each of these
fw-update-strategy class objects' 'from_dict' function.

This update also removes another object member, this one
currently unused, that would also not be de-serialized
if it were to be put to use as is in the future.

Test Plan:

PASS: Verify end-to-end orchestrated fw update (x2)

Closes-Bug: 1929251
Change-Id: I4540d1712f4dfee74e592c4f3ebce9c7cc913ab2
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>