Specify the nodeset for zuul jobs
The py2.7 jobs need to specify xenial
Changed py37 to py36 and specify bionic.

The un-specified python3 jobs work fine on either
focal or bionic.

zuul is not setup to trigger off code changes in this repo
so no source code changes are required to trigger the zuul
jobs

Partial-Bug: 1928978
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: Iab9c8727a0f16fa7ff02c20ca3bec5622abe7bd7