py3: nfv update range compatibility
Range returns a list on python2 and a range object on python3.

Convert to list where range is not iterated over.

This commit is split from Charles Short's work
Iecfe9b83245847b07aaaa0ec1aadad4734a9d118 and drops the usage
of the six library for six.moves.range as it is not needed.

Story: 2008454
Task: 42727
Co-authored-by: Charles Short <charles.short@windriver.com>
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: I97ca43b258e89ccd5f94ef17c768e5080fc946af