Kube rootca update orchestration integration
Updates to orchestration based on recent sysinv commits.

"Kube rootca update abort - API"
 -  changed the URL from /upload to /upload_cert
 - introduced a new state which requires the vim state
 machine to be updated to support resume from that state
 (it goes from abort to start)
 - required changes to how the the 'complete' step was invoked .

"CLIs for kube rootca update procedure"
-  changed the values for the states used by kube rootca
 update orchestration.

Improvements:
 - The subject and expiry_date validation is duplicated in the
VIM so the strategy does not need to be run to see if the inputs
are valid.
Note: if the strategy is created with a valid expiry date, and
not applied until after the date becomes invalid, that will not
be caught until that step is processed.

 - The Client was converting the error strings to lower case before
displaying them.  This made it very difficult to determine the
expected fields for things like certificate subject.

 - upload_cert option for the VIM now works.
Note: the path to the cert should be accessible from both controllers
otherwise it may fail to upload after a SWACT.

Story: 2008675
Task: 43131
Depends-On: https://review.opendev.org/c/starlingx/config/+/805375
Depends-On: https://review.opendev.org/c/starlingx/config/+/805878
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: Ia27b65ba5142516d5a62c5225c8498997367fd6e