Upgrade orchestration handle host-unlock after HostUpgrade
During a platform upgrade, the HostUpgrade step results in a
host-reinstall which requires the inventory to be repopulated.
There is a potential sysinv raised exception which may occur
when attempted too early.

Similar was done for AIO/workers however, another condition
requiring retry may also occur on controller/storage due to
the need to repopulate the inventory prior to the host-unlock.

Test Plan:
Perform Orchestrated upgrade of Controller Hosts
Perform Orchestrated upgrade of Storage Hosts

Closes-Bug: 1946255
Signed-off-by: John Kung <john.kung@windriver.com>
Change-Id: I59db9d34cae8331bd64135316b931bdf9aec051a