Kube Root ca update orchestration must ignore cert alarms
The 'Expiring' and 'Expired' alarms for the certificate need
to be added to the ignored alarms list when creating an update
strategy for kube rootca update orchestration.

The expiring alarm is often used to indicate that an orchestration
is needed.
The expired alarm is added for completeness, even though there is
no automated recovery from it.

Test Plan:
  Tested strategy can be created/applied when expiring alarm reported.

Closes-Bug: 1948673
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: I5348a0011f52db7bf9872ba0de13a8802688929f