Propagate unexpected errors from nfv client
The code that was iterating over unexpected errors would
accidentally 'return' rather than raise the error during
the area where it extracts a reason from the exception.

This meant that if the API returned an internal error (500)
the code would mistakenly treat the result as a None response.

This was particularly troublesome when performing GET operations.

Now it reports the error and allows the exception to be raised.

Test Plan:
  PASS: Verify the CLI reports an error when invoked while the API
is beng restarted.
  PASS: Verify the CLI returns a valid value when invoked under
normal circumstances.
  PASS: Verify the API reports an error rather than None when a
strategy is queried during an API restart.

Closes-Bug: 1950207
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: I7dc62b28425f09edabf2e6e4914aebc013da3cf4