Setting PYTHON3=no for LOCI images
Recently, LOCI changed its default so images are python3 if nothing is
specified at the *docker_image files. With the recent upversion of LOCI
in the build tools [1], we need to explictly set PYTHON3=no to images
that are still using python2. This commit fixes the StarlingX images to
use this parameter

TEST PLAN
PASS build the stx-nova-api-proxy image on the stable branch

[1] https://opendev.org/starlingx/root/commit/00bd632a68b8172f5d4baec2b1e4f8a08b9d5be5

Closes-Bug: #1958696

Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: I3b0a1e1d4b64527ce03192fb0f38bb7eba0d05c0