NFV VIM API request audit logging
Here is added a hook to the WSGI server that receives the NFV VIM API
requests. This hook logs the needed request data to
"/var/log/nfv-vim-api.log". All requests are logged except the ones of
"GET" type.

The code is a port from the same hook that exists in "starlingx/config"
repository, also called "AuditLogging". In this case, the logging
infrastructure present in the repository was used instead of oslo log.

Note: there are no "PATCH" or "PUT" requests present in NFV VIM API v1.

Test Plan:

PASS: Successfully deploy an AIO-SX and verify that the logs of
"nfv-vim-api" program are present in file "/var/log/nfv-vim-api.log".
PASS: In the deployed AIO-SX, execute command "sw-manager patch-strategy
show" and check that no GET request was logged in "nfv-vim-api.log".
PASS: In the deployed AIO-SX, execute command "sw-manager patch-strategy
create" and check that a "POST /api/orchestration/sw-patch/strategy"
request was logged in "nfv-vim-api.log" with the correct request fields.
PASS: In the deployed AIO-SX, execute command "sw-manager patch-strategy
apply" and check that a "POST
/api/orchestration/sw-patch/strategy/actions" request was logged in
"nfv-vim-api.log" with the correct request fields.
PASS: In the deployed AIO-SX, execute command "sw-manager patch-strategy
delete" and check that a "DELETE /api/orchestration/sw-patch/strategy"
request was logged in "nfv-vim-api.log" with the correct request fields.
PASS: In the deployed AIO-SX, execute command "sw-manager
upgrade-strategy create" and check that a "POST
/api/orchestration/sw-upgrade/strategy" request was logged in
"nfv-vim-api.log" with the correct request fields.
PASS: In the deployed AIO-SX, execute command "sw-manager
upgrade-strategy apply" and check that a "POST
/api/orchestration/sw-upgrade/strategy/actions" request was logged in
"nfv-vim-api.log" with the correct request fields.
PASS: In the deployed AIO-SX, execute command "sw-manager
upgrade-strategy delete" and check that a "DELETE
/api/orchestration/sw-upgrade/strategy" request was logged in
"nfv-vim-api.log" with the correct request fields.
PASS: In the deployed AIO-SX, execute command "sw-manager
fw-update-strategy create" and check that a "POST
/api/orchestration/fw-update/strategy" request was logged in
"nfv-vim-api.log" with the correct request fields.
PASS: In the deployed AIO-SX, execute command "sw-manager
fw-update-strategy apply" and check that a "POST
/api/orchestration/fw-update/strategy/actions" request was logged in
"nfv-vim-api.log" with the correct request fields.
PASS: In the deployed AIO-SX, execute command "sw-manager
fw-update-strategy delete" and check that a "DELETE
/api/orchestration/fw-update/strategy" request was logged in
"nfv-vim-api.log" with the correct request fields.
PASS: Successfully build package "nfv" of this repository as Debian
package.

Story: 2009824
Task: 44614

Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: I6becd2e4429cf5a98f4197422694406970d6bbc4