Add missing headers in NFV VIM API requests
The headers 'X-User', 'X-Tenant' and 'X-User-Domain-Name' were added
to NFV VIM API requests and are logged when received by the API server.

Test Plan:

PASS: Successfully deploy a Distributed Cloud (1 subcloud) using an
image with this commit present and verify in the contents of
"/var/log/nfv-vim-api.log" that the NFV VIM API server started
successfully in the central cloud and in the subcloud.
PASS: In the deployed subcloud, execute command
"sw-manager patch-strategy show" and check that the command was
successful.
PASS: In the deployed subcloud, execute each of the following commands
and check that the log of the received requests are present in
"/var/log/nfv-vim-api.log" with the fields "user", "tenant" and
"domain" correctly filled:
sw-manager patch-strategy apply
sw-manager patch-strategy abort
sw-manager patch-strategy create
sw-manager patch-strategy delete
PASS: In the deployed central cloud, create and execute a patch-strategy
using dcmanager CLI to verify the compatibility with the changes.
PASS: In the deployed central cloud, create and execute a patch-strategy
using Horizon interface to verify the compatibility with the changes.
PASS: Successfully build package "nfv" as Debian package.

Story: 2009824
Task: 44795

Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: Iccbc54440d3f62da46ccf8991fb075f14a9562c2