Debian: Eliminate nfv IOError logs during api calls
The underlying file object for sockets is different on
python2 (Centos) and python3 (Debian).

 The python2 socket._fileobject appears to allow 'flush'
to be invoked on it while closed.

 The python3 SocketWriter object does not appear to
permit operations like flush after it has been closed
and raises the IOError.

The sequence of calls is the same in both
environments, but the underlying issue is that a 'done'
method is being invoked twice for some API calls.

This change removes an extra call to 'done' by having
it just log a deprecation debug entry.

There should be no difference in functionality on either
environment, and this also stops the error logs on Debian.

Test Plan:
  Verified there are no errors in nfv logs on Centos.
  Verified there are no errors in nfv logs on Debian.
  Verified the sysinv logs are clean on both platforms
when invoking these particular nfv api calls.

Closes-Bug: #1967592
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I41233aabbebd8d6162143e01981170e691c7011d