Increase timeout for networking step on k8s upgrade
Kubernetes upgrade might fail during the Upgrade Networking
Step with timeout message when upgrading subclouds. The default
timeout of 600s from the parent class does not seem to be enough
for some subclouds to download the networking images.

This commit increases the timeout of the Upgrade Networking Step.

Test Plan:
PASS: upgrade k8s version with 90 subclouds in parallel

Closes-bug: 1973781
Change-Id: I686b9582daf14f0520fc0c2cb7100816c372107e
Signed-off-by: Heitor Matsui <HeitorVieira.Matsui@windriver.com>