Debian: Fix ugly error deleting a patch strategy
A JSONDecodeError is raised when deleting a patch strategy
that exists. (the error is not seen if no strategy exists)

sw-manager patch-strategy delete
Expecting value: line 1 column 1 (char 0)

The issue is the JSON decoder does not handle an empty
string. The empty string is a byte string on debian which
runs on python3 so the comparison check needed to be
updated.

Test Plan:
  Verify that a patch strategy can be deleted on Debian
without ugly messages.
  Verify that a non existant patch strategy can be deleted
on Debian without ugly messages.

Partial-Bug: 1974475
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I18f84e20abf43a342e996c4752654bad3261e5a6