Fix tox-docs failing sphinx
A new version of sphinx was released May 29 2022
which requires a language setting in config otherwise
a warning (treated as error) causes the sphinx operation
to fail.

Updating a file header to trigger zuul.
Updated the sphinx config file to correct the issue.

The sphinx behavioural change is mentioned here:
https://github.com/sphinx-doc/sphinx/issues/10062
https://github.com/sphinx-doc/sphinx/issues/10474

Note: The bug referenced below is for the config repo,
      but it affects multiple StarlingX repositories.

Closes-Bug: 1976377
Change-Id: I7bb7886f27ce6190f8e23270efc1aadb7364fc90
Signed-off-by: Heitor Matsui <HeitorVieira.Matsui@windriver.com>