Detect and kill hung vim process
There is a specific scenario when vim starts where threads keep
stalled and then after 600s the subsystems are marked as not sane.
In this situation vim is supposed to quit, but instead it stays
alive but in a failed state.

This commit adds an extra step in vim monitoring ocf script to
test if vim rpc port is open and responding, and if not, vim
is forcefully killed.

Test Plan
PASS: install/bootstrap/unlock AIO-SX
PASS: get vim to a not-sane state and observe it being killed
      and restarted in this scenario

Closes-bug: 1979992
Change-Id: I55777a0858507634a362e58b9da281492809a03d
Signed-off-by: Heitor Matsui <HeitorVieira.Matsui@windriver.com>