Increase sysinv API timeout from 45 to 60 seconds
When the VIM issues a sysinv command, it will terminate the thread
and report a 'timed out' for the plugin activity if the command
takes longer than 45 seconds.

Under heavy load (10 hosts patching in parallel) sysinv can take longer
than 45 seconds to generate the hieradata for an unlock, and therefore
the patch strategy can fail.

This fix bumps the value up to 60.

Any future improvements will focus on speeding up the generation
of sysinv hieradata.

Closes-Bug: 1986972
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I1dbb6f1e3c5529de9199dea8453f32e954da2f51