Fix host-delete for host that has k8s pods
Attempting to delete a host that had pods on it
was failing.  This was due to an exception in nfv
when calling the kubernetes client.

On CentOS 7 python-kubernetes version 8 is used.
On Debian Bullseye python-kubernetes version 12 is used.

The delete_node API has changed in version 12.
'body' needs to be a named argument.

Closes-Bug: 1995742
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I1ad56f4363ed2cc416335f37fdf20b39cd22ee1b