Unable to resolve v3 url when sw-manager is invoked
When the OpenStack RC file is exported from Horizon
the url in the OS_AUTH_URL variable does not include the /v3
API version. This causes that when the sw-manager command
is invoked and trying to get the token, it returns error 300.
This fix is to catch the error and get the URL of the location
which points to the /v3 API version.

Closes-Bug: 1998949

Test Plan:

- Download OpenStack RC file from horizon then source it in
remote machine.
- Use remote-cli to execute sw-manager patch-strategy show

Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>
Change-Id: Ib5112bed26c2b773a460377af8be0c7ec12c80ae