Deactivate guest related services
The guest services are currently not being used since we
went to containerized openstack. Since the service is no
longer being used and it is currently causing coredump
issues [1] in debian environment we are deactivating all
related guest services. The service can be reactivated in
the future if needed (we created a Storyboard [2] to keep
track of this modification).

[1] https://bugs.launchpad.net/starlingx/+bug/2000168
[2] https://storyboard.openstack.org/#!/story/2010520

Test Plan:
PASS: Generate the debian image without the code that
adds all the guest related services.
PASS: Perform a install with the created debian image
and check that no guest related services are being
installed.

Partial-Bug: 2000168

Signed-off-by: Rafael Falcao <rafael.vieirafalcao@windriver.com>
Change-Id: I55ed75a7dc02cb517b1343562c84cbaa5c685dd6