Remove HostTask actions of guest related services
Since we are currently deactivating all guest
related services [1][2][3] we now need to remove
the add, delete, enable and disable sections of
those services from the HostTask state machine
to be able to correctly perform tasks like
lock and unlock of the host.

[1] https://review.opendev.org/c/starlingx/stx-puppet/+/869474
[2] https://review.opendev.org/c/starlingx/tools/+/870433
[3] https://review.opendev.org/c/starlingx/nfv/+/869817

Test Plan:
PASS: Generate the debian image without the code that enable
and disable the guest services during a lock/unlock
PASS: Perform lock/unlock of all controllers in the system
with stx-openstack applied.
PASS: Perform a delete/add of a host in the system with
stx-openstack applied.

Closes-Bug: 2000483

Signed-off-by: Rafael Falcao <rafael.vieirafalcao@windriver.com>
Change-Id: Iaa2d0ff71a21fb6ee03ecaa53415859982070985