Add support for SQLAlchemy 1.4 to NFV
An internal variable was dropped in SQLALchemy 1.4.

This update attempts to use the newer syntax from 1.4
and falls back to the older syntax if the newer mechanism
cannot be used.

This will allow StarlingX to migrate to a newer version of
SQLALchemy without breaking the NFV database code.

Test Plan:
  PASS: verify coverage using SQLAlchemy 1.2 that the
 exception code is executed and unit tests pass.
  PASS: verify coverage using SQLAlchemy 1.4 that the
 exception code is not executed and unit tests pass.
  PASS: Build/Install/Bootstrap/Unlock AIO-SX to verify
 existing runtime behaviour is not impacted.

Story: 2010531
Task: 47237

Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I4063dac0b3229b4c1fdb6c5121154665ffc32903