Replace mock with unittest.mock
mock module has been deprecated in favor of unittest.mock
which is part of the standard python library.

This review removes mock as a third party dependency, and
updates all unittest imports to use unittest.mock instead.

nova-api-proxy had mock as a test requirement, however it
was not using it.

This change has no impact on runtime code.

Test Plan:
  PASS: tox passing for nfv
  PASS: build-pkgs -p nfv
  PASS: tox passing for nova-api-proxy
  PASS: build-pkgs -p nova-api-proxy

Story: 2010531
Task: 47142

Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I991d07e87fbdc7fecc4e790550df2472b9b5b24b