Convert yaml.load to yaml.safe_load
yaml.load is considered unsafe. yaml.safe_load is the
recommended method to use.

The yaml.load is only being called from scenario tests,
and from a forensic tool that is unused. Therefore this
code change has no runtime impact.

This allow unsuppressing the following pylint error
 E1120 no-value-for-parameter

This allow unsuppressing the following bandit error
 B506: Test for use of yaml load

Adding a new bandit zuul job, and removing the pinning
of the version of bandit for this repo.

Test Plan:
  PASS: tox -e pylint
  PASS: tox -e bandit
  PASS: build-pkgs -p nfv

Story: 2010531
Task: 47174

Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: Ie60dcf98425c95e103d22e7be4212de1954550bf