Replace SafeConfigParser with ConfigParser
'SafeConfigParser' is a deprecated class.  Replace the
reference to it with its non-deprecated counterpart
called 'ConfigParser'

This change only affects how the debug.ini file is loaded,
which is a file used for setting log levels within NFV.

The change is not meant to affect any runtime behaviour.
The intention of this code change is for this to behave
the same as using SafeConfigParser.

pylint was reporting this as a deprecated class (W4904)

Test Plan:
  PASS: added a new unit test to verify coverage
  PASS: build ISO and unlock AIO-SX and verify vim working
   and startup logs appear unchanged

Story: 2010531
Task: 47215

Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: Id0c22f9847489b171c28c11f8d8215902d444241