Debian: Fix nova actions
Since the platform migration to Debian, it was observed that the
following Nova actions stopped working:
  - pause;
  - unpause;
  - suspend;
  - resume;
  - live-migration.

The reason behind that is that some packages related to Nova, which have
already been migrated to Debian, still have some incompatibilities with
Python 3. Consequently, whenever these Nova actions were executed, some
exceptions occurred on the nova-api-proxy and NFV side, preventing them
from working.

Therefore, this change aims to improve this compatibility.

Most of the changes were necessary due to the fact that in Python 3
there is more of a distinction between `bytes` and `str`, whereas in
Python 2 `bytes` is just an alias for `str`.

Test Plan (on AIO-DX):
PASS - Successfully perform a VM pause, unpause, suspend, resume.
PASS - Successfully perform a VM live-migration.

Closes-Bug: 2003813

Signed-off-by: Luan Nunes Utimura <LuanNunes.Utimura@windriver.com>
Change-Id: I918fe6e3deaa68630c797449649012e9fbf16fe4